+++ hbnb/arm/custom-settings/config/root/etc/init.d/build.sh
@@ -9,24 +9,16 @@
 # X-Interactive:     true
 # Short-Description: build HotspotBnB
 ### END INIT INFO
-
 #custom-settings/config/files/root/etc/init.d to raspberrypi-ua-netinst/config/files/root/etc/init.d
-sudo echo "v0.0.1-rc.9-hbnb-arm.0.8 BUILD ..." &&
+sudo echo "v0.0.1-rc.9 BUILD ..." &&
 sleep 25 &&
-printf "%s" "pinging 8.8.8.8 to be sure ..."
-while ! timeout 0.2 ping -c 1 -n 8.8.8.8 &> /dev/null
-do
-    printf "%c" "."
-done
-printf "\n%s\n"  "ping success. proceeding ..." &&
-
+sudo apt update -y &&
 sudo apt install -y net-tools git git-svn subversion apache2 &&
-
 sudo echo "Fetching the Hotspotβnβ Dashboard ..." &&
 sudo sleep 2 &&
 sudo printf "\n%s\n"  "Executing Main Method ..." &&
 sudo mkdir -p /tmp/html/ &&
-svn co --depth infinity https://github.com/unclehowell/datro/trunk/static/gui/  \
+sudo svn co --depth infinity https://github.com/unclehowell/datro/trunk/static/gui/  \
                         /tmp/html/ &&
 sleep 5 &&
 sudo cp -r /tmp/html/* /var/www/html >&- 2>&- &&
@@ -34,19 +26,20 @@ sudo sleep 0.1 &&
 sudo rm -r /tmp/html &&
 
 sleep 5 &&
-sudo systemctl restart apache2 &&
+sudo chown root:root /etc/sudoers /etc/sudoers.d -R && #fix for error (pid 1000 when it should be 0)
+sudo echo "www-data ALL = NOPASSWD: /sbin/reboot, /sbin/halt" >> /etc/sudoers &&
+sudo chmod -R 755 /var/www/html &&
 
-echo "www-data ALL = NOPASSWD: /sbin/reboot, /sbin/halt" >> /etc/sudoers &&
 sudo systemctl restart apache2 &&
-sudo echo "Setting up www-data user & permissions ..."
-sudo grep www-data /etc/passwd &&
-sudo grep www-data /etc/group &&
-sudo useradd -g www-data -s /usr/sbin/nologin -m -d /home/www-data www-data &&
-sudo usermod -a -G www-data pi &&
-sudo systemctl restart apache2 &&
+#sudo echo "Setting up www-data user & permissions ..."
+#sudo grep www-data /etc/passwd &&
+#sudo grep www-data /etc/group &&
+#sudo useradd -g www-data -s /usr/sbin/nologin -m -d /home/www-data www-data &&
+#sudo usermod -a -G www-data pi &&
+#sudo systemctl restart apache2 &&
 
 if [ ! -d "/var/www/html/" ]; then
-    echo "The Hotspotβnβ Dashboard failed to download - Rebooting and trying again"
+    echo "Hotspotβnβ ran into an issue installing - reboot or begin again"
     sleep 2 &&
     reboot
 else